adjusting prime select behaviour to include option to reboot later

+++ src/ng-app/app/prime-dialog/prime-dialog.component.ts
@@ -1,3 +1,22 @@
+/*!
+ * Copyright (c) 2019-2023 TUXEDO Computers GmbH <tux@tuxedocomputers.com>
+ *
+ * This file is part of TUXEDO Control Center.
+ *
+ * TUXEDO Control Center is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * TUXEDO Control Center is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with TUXEDO Control Center.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 import { Component, OnInit } from "@angular/core";
 import { UtilsService } from "../utils.service";
 import { ElectronService } from "ngx-electron";
@@ -12,6 +31,8 @@ export class PrimeDialogComponent implements OnInit {
     primeSelectMode: string;
     loadingBar = false;
     langId: string;
+
+    dialogStatus: string = "info";
 
     constructor(
         private electron: ElectronService,
@@ -35,27 +56,28 @@ export class PrimeDialogComponent implements OnInit {
         this.langId = this.utils.getCurrentLanguageId();
     }
 
-    public async applyPrimeConfig() {
+    public setDialogStatus(status: string) {
+        this.dialogStatus = status;
+    }
+
+    public async applyPrimeConfig(rebootStatus: string) {
+        this.setDialogStatus("loading");
+
         this.loadingBar = true;
         const status = await this.config.pkexecSetPrimeSelectAsync(
             this.primeSelectMode
         );
+
         if (status) {
-            this.utils.execCmd("reboot");
+            console.log("rebootStatus: ", rebootStatus);
+            if (rebootStatus === "REBOOT") {
+                this.utils.execCmd("reboot");
+            }
+            this.electron.ipcRenderer.send("prime-window-close");
         }
         if (!status) {
             this.electron.ipcRenderer.send("prime-window-close");
         }
-    }
-
-    public openHelpPage() {
-        const helpLinks = {
-            en: "https://www.tuxedocomputers.com/en/PRIME-GPU-Render-Offloading/GPU-on-demand-Mode-Guide.tuxedo",
-            de: "https://www.tuxedocomputers.com/de/Der-umfassende-PRIME-GPU-Render-Offloading/GPU-on-demand-Mode-Leitfaden.tuxedo",
-        };
-        this.electron.shell.openExternal(
-            helpLinks[this.langId] || helpLinks.en
-        );
     }
 
     public closeWindow() {
